REFACT: shader를 담당하는 코드 분리
